Separate counting methods in files for easy manipulation
In each file, each line number corresponds to sura number.
This makes it easy to find any sura aya count by jumping to its line.
It also makes git commits easy to understand as the changed sura will
be the only one highlighted in diff, not the whole one-line array.